🚀 4단계 - 수강신청(요구사항 변경)

+++ src/test/java/nextstep/session/infrastructure/StudentRepositoryTest.java
@@ -28,11 +28,11 @@ class StudentRepositoryTest {
     @Test
     void crud() {
         Student student = new Student(JAVAJIGI, createSession());
-        int savedId = studentRepository.save(student);
+        Long savedId = studentRepository.save(student);
 
         assertThat(savedId).isEqualTo(1);
 
-        Student savedStudent = studentRepository.findById((long) savedId).orElse(null);
+        Student savedStudent = studentRepository.findById(savedId).orElse(null);
 
         assertThat(student.getSessionId()).isEqualTo(savedStudent.getSessionId());
     }
